create NeuralNetworkTest to test instrumentation

+++ src/main/java/com/cluttered/cryptocurrency/ann/NeuralNetwork.java
@@ -9,10 +9,10 @@ import java.util.List;
  */
 public class NeuralNetwork {
 
-    private final long inputSize;
+    private final Long inputSize;
     private final List<Layer> layers;
 
-    public NeuralNetwork(final int inputSize, final List<Layer> layers) {
+    public NeuralNetwork(final Long inputSize, final List<Layer> layers) {
         this.inputSize = inputSize;
         this.layers = layers;
     }
